# copiloto-function/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Copiloto Function Gateway",
    "version": "3.5",
    "description": "Sistema completo con memoria automática integrada, queries dinámicas, consulta previa y guardado automático de interacciones en Cosmos DB. Nota: los endpoints legacy /api/probar-endpoint e /api/invocar fueron retirados; usa /api/copiloto como router semántico.",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://copiloto-func.ngrok.app"
    },
    {
      "url": "https://copiloto-semantico-func-us2.azurewebsites.net"
    }
  ],
  "paths": {
...
    "/api/eliminar-archivo": {
      "post": {
        "summary": "Eliminar archivo",
        "operationId": "eliminarArchivo",
        "tags": [
          "Azure Storage"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ruta": {
                    "type": "string",
                    "description": "Ruta del archivo a eliminar"
                  }
                },
                "required": [
                  "ruta"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archivo eliminado correctamente"
          },
          "400": {
            "description": "Solicitud inválida"
          },
          "404": {
            "description": "Archivo no encontrado"
          }
        }
      }
    },
...
    "/api/sugerencias": {
      "get": {
        "summary": "💡 SUGERENCIAS CONTEXTUALES - Genera sugerencias basadas en historial",
        "description": "Genera sugerencias inteligentes basadas en el historial de interacciones de la sesión. Usa queries dinámicas para analizar patrones y recomendar acciones.",
        "operationId": "sugerencias",
        "tags": [
          "🧠 Memoria Integrada"
        ],
        "security": [],
        "parameters": [
          {
            "name": "Session-ID",
            "in": "header",
            "description": "ID de sesión para obtener sugerencias contextuales",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "description": "Número máximo de sugerencias",
            "schema": {
              "type": "integer",
              "default": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sugerencias generadas exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exito": {
                      "type": "boolean"
                    },
                    "sugerencias": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "basadas_en_historial": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}